+++ src/config.php
@@ -1,0 +1,9 @@
+<?php
+
+//DBs
+const HOST = '127.0.0.1';
+const LOGIN = 'root';
+const PASSWORD ='';
+const DBNAME = 'doingsdone';
+
+const USERID = 3;
